Front validate

+++ app/Http/Controllers/UserController.php
@@ -343,17 +343,9 @@ class UserController extends Controller
         }
 
         // Delete all accesses of the selected user
-        if (Adm::where('user_id', $id)->get()) {
-            Adm::where('user_id', $id)->delete();
-        }
-
-        if (Tec::where('user_id', $id)->get()) {
-            Tec::where('user_id', $id)->delete();
-        }
-
-        if (Sup::where('user_id', $id)->get()) {
-            Sup::where('user_id', $id)->delete();
-        }
+        if (Adm::where('user_id', $id)->get()) {Adm::where('user_id', $id)->delete();}
+        if (Tec::where('user_id', $id)->get()) {Tec::where('user_id', $id)->delete();}
+        if (Sup::where('user_id', $id)->get()) {Sup::where('user_id', $id)->delete();}
 
         // Delete the selected user
         $deleted = $this->user->where('id', $id)->delete();
